+++ changelog.md
@@ -8,6 +8,7 @@ Codion Change Log
 ### is.codion.framework.db
 - DefaultLocalEntityConnection, selecting single values of entities with custom query columns now supported.
 - EntityResultPacker now fetches values by column name when the entity query has custom columns, query column order no longer has to match the attribute definition order.
+- EntityConnection.Select.queryTimeout() renamed timeout().
 
 ## 0.18.44
 ### is.codion.common.core